+++ crates/impresspress-core/src/blocks/dev/retention.rs
@@ -78,7 +78,7 @@ pub const RETAINED_GENERATIONS: usize = 20;
 /// pinned blobs against the workspace quota forever and, past the cap, would
 /// start pushing the serving generation out of this set.
 pub async fn retained(ctx: &dyn Context) -> Result<Vec<GenerationRow>, WaferError> {
-    let mut rows = generations::list_recent(ctx, RETAINED_GENERATIONS as i64).await?;
+    let mut rows = generations::list_recent(ctx, RETAINED_GENERATIONS as u32).await?;
     if rows.len() < RETAINED_GENERATIONS {
         // The whole ledger is inside the window: the other two queries could
         // only return rows this one already holds.
@@ -111,7 +111,7 @@ pub async fn retained(ctx: &dyn Context) -> Result<Vec<GenerationRow>, WaferErro
 /// page is asked for, so an empty page means done and a ledger far past the
 /// window is collected in full rather than down to one page's worth of rows.
 pub async fn prune(ctx: &dyn Context) -> Result<Vec<GenerationRow>, WaferError> {
-    let newest = generations::list_recent(ctx, RETAINED_GENERATIONS as i64).await?;
+    let newest = generations::list_recent(ctx, RETAINED_GENERATIONS as u32).await?;
     let Some(boundary) = newest.get(RETAINED_GENERATIONS - 1) else {
         // Fewer rows than the window holds: nothing is outside it.
         return Ok(Vec::new());
